APIS-4714 - Extracted out more traits
