+++ test/uk/gov/hmrc/apidocumentation/controllers/AuthorisationControllerSpec.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.apidocumentation.controllers
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import uk.gov.hmrc.apidocumentation.views.html._
-import uk.gov.hmrc.apidocumentation.controllers.utils.NavigationServiceMock
+import uk.gov.hmrc.apidocumentation.controllers.utils.{NavigationServiceMock, PageRenderVerification}
 
 class AuthorisationControllerSpec extends CommonControllerBaseSpec with PageRenderVerification {
   trait Setup extends NavigationServiceMock {
